requirements.txt
aiohttp
aiohttp-retry
aiosignal
asgiref
async-timeout
attrs
certifi
cfgv
chardet
charset-normalizer
cloudinary==1.36.0
coreapi
coreschema
decouple
defusedxml
distlib
Django
django-allauth
django-cloudinary-storage==0.3.0
django-cors-headers
django-rest-auth
django-rest-swagger
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
drf-yasg
filelock
frozenlist
gunicorn
identify
idna
inflection
isort
itypes
Jinja2
MarkupSafe
multidict
nodeenv
oauthlib
openapi-codec
packaging
Pillow
platformdirs
pre-commit
PyJWT
python-dotenv
python3-openid
pytz
PyYAML
requests
requests-oauthlib
simplejson
six
sqlparse
tk
twilio
uritemplate
urllib3
virtualenv
whitenoise
yarl
fuzzywuzzy
pydub
google-auth
google-auth-httplib2
google-auth-oauthlib
oauthlib
django-allauth
django-rest-auth
